refactor: Updated to Swift 3 and Xcode 8
