--- Example/Podfile
@@ -1,12 +1,0 @@
-source 'https://github.com/CocoaPods/Specs.git'
-use_frameworks!
-
-target 'GaugeView_Example', :exclusive => true do
-  pod "GaugeView", :path => "../"
-end
-
-target 'GaugeView_Tests', :exclusive => true do
-  pod "GaugeView", :path => "../"
-
-  
-end
